Changes requested in code review.
